requirements.txt
azure-functions
./custom-modules/example_pkg_davpo-0.0.1-py3-none-any.whl